Crate queue

A simple and easy wrapper around Vec to implement a FIFO queue. This is no fancy, advanced data type but something simple you can use easily until or unless you need something different.

Example

use queue::Queue;

let mut q = Queue::new();

q.queue("hello").unwrap();
q.queue("out").unwrap();
q.queue("there!").unwrap();

while let Some(item) = q.dequeue() {
    println!("{}", item);
}

Outputs:

hello
out
there!

Structs

Queue
A first in, first out queue built around Vec.